added zone pricings management section

+++ db/migrate/20160618053411_create_zone_pricings.rb
@@ -2,7 +2,7 @@ class CreateZonePricings < ActiveRecord::Migration
   def change
     create_table :zone_pricings do |t|
       t.references :zone, index: true, foreign_key: true
-      t.integer :weight
+      t.decimal :weight
       t.decimal :price
 
       t.timestamps null: false
